Fixed space_added_updated
When a new hackerspace gets added, the previously entered
hackerspace will be "updated", because a comma gets appended
to the entry.
This patch should show the actually changed hackerspace in
the commit message (instead of the unchanged previously
entered space). It should also work if an existing hackerspace
is changed.

I did not test this!